+++ src/spatRaster.h
@@ -1049,13 +1049,13 @@ class SpatRaster {
 		SpatRaster weighted_mean(SpatRaster w, bool narm, SpatOptions &opt);
 		SpatRaster weighted_mean(std::vector<double> w, bool narm, SpatOptions &opt);
 
-		SpatRaster warper(SpatRaster x, std::string crs, std::string method, bool mask, bool align, bool resample, std::string pipeline, std::vector<double> AOI, double desired_accuracy, bool allow_ballpark, double xscale, double yscale, std::vector<std::string> warp_opts, std::vector<std::string> trans_opts, SpatOptions &opt);
+		SpatRaster warper(SpatRaster x, std::string crs, std::string method, bool mask, bool align, bool resample, std::string pipeline, double xscale, double yscale, std::vector<std::string> warp_opts, std::vector<std::string> trans_opts, SpatOptions &opt);
 		SpatRaster warper(SpatRaster x, std::string crs, std::string method, bool mask, bool align, bool resample, SpatOptions &opt) {
-			return warper(x, crs, method, mask, align, resample, "", std::vector<double>(), -1.0, true, 0, 0, std::vector<std::string>(), std::vector<std::string>(), opt);
+			return warper(x, crs, method, mask, align, resample, "", 0, 0, std::vector<std::string>(), std::vector<std::string>(), opt);
 		}
-		SpatRaster warper_by_util(SpatRaster x, std::string crs, std::string method, bool mask, bool align, bool resample, std::string pipeline, std::vector<double> AOI, double desired_accuracy, bool allow_ballpark, double xscale, double yscale, std::vector<std::string> warp_opts, std::vector<std::string> trans_opts, SpatOptions &opt);
+		SpatRaster warper_by_util(SpatRaster x, std::string crs, std::string method, bool mask, bool align, bool resample, std::string pipeline, double xscale, double yscale, std::vector<std::string> warp_opts, std::vector<std::string> trans_opts, SpatOptions &opt);
 		SpatRaster warper_by_util(SpatRaster x, std::string crs, std::string method, bool mask, bool align, bool resample, SpatOptions &opt) {
-			return warper_by_util(x, crs, method, mask, align, resample, "", std::vector<double>(), -1.0, true, 0, 0, std::vector<std::string>(), std::vector<std::string>(), opt);
+			return warper_by_util(x, crs, method, mask, align, resample, "", 0, 0, std::vector<std::string>(), std::vector<std::string>(), opt);
 		}
 
 		std::vector<double> warp_scale(SpatRaster x, size_t n);